+++ example/lib/data_table2_tests.dart
@@ -1,5 +1,5 @@
 import 'package:data_table_2/data_table_2.dart';
-import 'package:data_table_2/paginated_data_table_2.dart';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';