+++ lib/src/async_paginated_data_table_2.dart
@@ -1,10 +1,17 @@
-import 'package:data_table_2/paginated_data_table_2.dart';
-import 'package:flutter/material.dart';
-
-import 'data_table_2.dart';
+part of 'paginated_data_table_2.dart';
 
 enum SourceState { none, ok, loading, error }
 
+class AsyncRowsResponse {
+  AsyncRowsResponse(this.totalRows, this.rows);
+  final int totalRows;
+  final List<DataRow> rows;
+}
+
+/// Implement this class and use it in conjunction with [AsyncPaginatedDataTable2]
+/// to allow asynchronous data fetching.
+/// Please overide [AsyncDataTableSource.getRows] and [DataTableSource.selectedRowCount]
+/// to make it legible as a data source.
 abstract class AsyncDataTableSource extends DataTableSource {
   SourceState _state = SourceState.none;
   SourceState get state => _state;
@@ -13,18 +20,28 @@ abstract class AsyncDataTableSource extends DataTableSource {
   Object? get error => _error;
 
   List<DataRow> _rows = [];
-  List<DataRow> get rows => _rows;
+  //List<DataRow> get rows => _rows;
+  int _totalRows = 0;
 
-  Future<List<DataRow>> getRows(int start, int end);
+  /// Override this method to allow the data source asynchronously
+  /// fetch data (e.g. from a server) and convert them to [DataRow]/[DataRow2]
+  /// entities consumed by [AsyncPaginatedDataTable2] widget.
+  /// Note that besides rows this method is also supposed to return
+  /// the total number of available rows (both values are packed into [AsyncRowsResponse] instance
+  /// returned from this method)
+  Future<AsyncRowsResponse> getRows(int start, int end);
 
   Future _getRows(int start, int end) async {
     _state = SourceState.loading;
     await Future(() => notifyListeners());
 
     try {
-      _rows = await getRows(start, end);
+      var data = await getRows(start, end);
+      _rows = data.rows;
+      _totalRows = data.totalRows;
     } catch (e) {
       _rows = [];
+      _totalRows = 0;
       _state = SourceState.error;
       _error = e;
       notifyListeners();
@@ -36,6 +53,19 @@ abstract class AsyncDataTableSource extends DataTableSource {
     notifyListeners();
     return;
   }
+
+  @override
+  DataRow? getRow(int index) {
+    if (_rows.length - 1 <= index) return _rows[index];
+
+    return null;
+  }
+
+  @override
+  int get rowCount => _totalRows;
+
+  @override
+  bool get isRowCountApproximate => false;
 }
 
 class AsyncPaginatedDataTable2 extends PaginatedDataTable2 {
@@ -61,7 +91,7 @@ class AsyncPaginatedDataTable2State extends PaginatedDataTable2State {
   Widget build(BuildContext context) {
     var source = widget.source as AsyncDataTableSource;
     if (source.state == SourceState.none) {
-      source._getRows(firstRowIndex, widget.rowsPerPage);
+      source._getRows(_firstRowIndex, widget.rowsPerPage);
       return SizedBox();
     } else if (source.state == SourceState.loading) {
       return Center(child: Text('Loading'));