+++ test/custom_features_test.dart
@@ -1,7 +1,6 @@
 @TestOn('!chrome')
 
 import 'package:data_table_2/data_table_2.dart';
-import 'package:data_table_2/paginated_data_table_2.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'test_utils.dart';