+++ test/paginated_data_table2_test.dart
@@ -4,11 +4,11 @@
 
 // Copyright 2021 Maxim Saplin - changes and modifications to original Flutter implementation of DataTable
 
-import 'package:data_table_2/paginated_data_table_2.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/gestures.dart' show DragStartBehavior;
+import 'package:data_table_2/data_table_2.dart';
 
 import 'test_utils.dart';