+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 !NOTE: the first 2 letters follow the stable version of FLutter SDK upon which the package is based on (i.e. which sources are being copied and modified) - if there're new APIs introuduced to DataTable and PaginatedDataTable with newer SDK relases those changes might not be present in the code provided by this package's widgets until the major and minor versions codes of this package match. The 3rd letter is the version of the package itself (most likely this part to be changed when bug fixes are intoduced or minor features are rolled out).  
 
+## 2.1.2
+- Asycnhronous data fecthing model via AsyncDataTableSource and tailored widget AsyncPaginatedDataTable2
+- Change of package exports (no need to import paginated_data_table_2.dart, data_table_2.dart is enough)
+
 ## 2.1.1
 - PaginatorController that allows externally control PaginatedDataTable2 state (e.g. switch pages, change page size etc.)
 - Custom paginator example for PaginatedDataTable2